chore: move watcher to s3uploader
